# assets/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Oficina API
  version: 1.0.0
  description: |-
    API da Oficina (Fase 1).

    Utilize `POST /auth/login` para obter um token JWT e, em seguida, informe `Bearer <token>` no cabeçalho `Authorization` ao chamar as rotas protegidas. No Swagger UI, use o botão **Authorize** para registrar o token.
servers:
  - url: http://localhost:8080
    description: Ambiente local (PORT=8080)
security:
  - bearerAuth: []
...
paths:
...
  /admin/services:
    get:
      tags: [Services]
      summary: Listar serviços
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Lista de serviços.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Services]
      summary: Criar serviço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreateRequest'
      responses:
        '201':
          description: Serviço criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          description: Corpo JSON inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid JSON body
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          description: Serviço duplicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "409"
                message: service with this name already exists
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
  /admin/services/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceIdParam'
    get:
      tags: [Services]
      summary: Obter serviço por ID
      responses:
        '200':
          description: Serviço encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          description: ID inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid ID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Serviço não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: service not found
    patch:
      tags: [Services]
      summary: Atualizar serviço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '204':
          description: Serviço atualizado.
        '400':
          description: ID inválido ou corpo JSON inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidId:
                  value:
                    code: "400"
                    message: invalid ID
                invalidJson:
                  value:
                    code: "400"
                    message: invalid JSON body
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Serviço não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: service not found
        '409':
          description: Serviço duplicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "409"
                message: service with this name already exists
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
    delete:
      tags: [Services]
      summary: Remover serviço
      responses:
        '204':
          description: Serviço removido.
        '400':
          description: ID inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid ID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Serviço não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: service not found
...
  /admin/customers/{customerId}/vehicles:
    parameters:
      - $ref: '#/components/parameters/CustomerIdParam'
    get:
      tags: [Vehicles]
      summary: Listar veículos por cliente
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Veículos do cliente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        '400':
          description: ID de cliente inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid ID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /admin/vehicles:
    get:
      tags: [Vehicles]
      summary: Listar veículos
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Lista de veículos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags: [Vehicles]
      summary: Criar veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateRequest'
      responses:
        '201':
          description: Veículo criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          description: Corpo JSON inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid JSON body
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '409':
          description: Placa duplicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "409"
                message: vehicle plate already exists
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
  /admin/vehicles/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceIdParam'
    get:
      tags: [Vehicles]
      summary: Obter veículo por ID
      responses:
        '200':
          description: Veículo encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: ID inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid ID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Veículo não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: vehicle not found
    patch:
      tags: [Vehicles]
      summary: Atualizar veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateRequest'
      responses:
        '204':
          description: Veículo atualizado.
        '400':
          description: ID inválido ou corpo JSON inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidId:
                  value:
                    code: "400"
                    message: invalid ID
                invalidJson:
                  value:
                    code: "400"
                    message: invalid JSON body
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Veículo não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: vehicle not found
        '409':
          description: Placa duplicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "409"
                message: vehicle plate already exists
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
    delete:
      tags: [Vehicles]
      summary: Remover veículo
      responses:
        '204':
          description: Veículo removido.
        '400':
          description: ID inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "400"
                message: invalid ID
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Veículo não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "404"
                message: vehicle not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Número máximo de registros a retornar.
      schema:
        type: integer
        minimum: 1
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Quantidade de registros a ignorar antes de retornar resultados.
      schema:
        type: integer
        minimum: 0
        default: 0
    ResourceIdParam:
      name: id
      in: path
      required: true
      description: Identificador do recurso.
      schema:
        type: string
        format: uuid
    CustomerIdParam:
      name: customerId
      in: path
      required: true
      description: Identificador do cliente.
      schema:
        type: string
        format: uuid
  responses:
    UnauthorizedResponse:
      description: Token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              value:
                code: "401"
                message: missing authorization token
            invalidToken:
              value:
                code: "401"
                message: invalid token
            expiredToken:
              value:
                code: "401"
                message: token expired
...
    UnprocessableEntityResponse:
      description: Entidade não processável (erros de validação ou regras de negócio).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_JSON
        message:
          type: string
          example: invalid JSON body
    IdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
          description: ID do recurso criado.
    ValidationError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_BODY
        message:
          type: string
          example: Body validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: password
              message:
                type: string
                example: Password is a required field
...
    Service:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: integer
          format: int64
          description: Preço do serviço em centavos (BRL).
    ServiceCreateRequest:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          minLength: 3
        price:
          type: integer
          format: int64
          minimum: 1
          description: Preço do serviço em centavos (BRL).
    ServiceUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
        price:
          type: integer
          format: int64
          minimum: 1
          description: Preço do serviço em centavos (BRL).
    ServiceListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
...
    Vehicle:
      type: object
      required: [id, customer_id, plate, brand, model, year]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
    VehicleCreateRequest:
      type: object
      required: [customer_id, plate, brand, model, year]
      properties:
        customer_id:
          type: string
          format: uuid
        plate:
          type: string
          description: Placa do veículo.
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
          minimum: 1900
    VehicleUpdateRequest:
      type: object
      properties:
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
          minimum: 1900
    VehicleListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
...